Updates Media selection to be cancelled upon navigation between Albums
Design team raised concern that having selected Media items in one Albums it was disconcerting to have the selection persisted when you switch to another Album. The primar reason was that the “Insert” button is displayed when Media is selected. THerefore you could have media “selected” in one Album whilst you were in another. Fixed this to clear Media selection upon folder change.